Rejigged the renderAnimPlay method and children for performance
It's now got a 'bones' branch and a 'static' branch, with the setting of the uniforms done as early as possible
This reduces the uniform seting code as well as a branch within the render loop

In valgrind I saw a small improvement in that method.
In real game testing I saw about a ~20 fps improvement for a game which had just started.