Fix compile against older version of FreeType
Older versions (including the bundled version for Windows users),
use int for bitmap.width and bitmap.rows; this was changed to
unsigned int later on.

Updated in FreeType commit b3500af7170 on 2014-11-21
http://git.savannah.gnu.org/cgit/freetype/freetype2.git/commit/?id=b3500af7170

From what I can tell, this means that 2.5.3 and earlier was using int,
and 2.5.4 and later was using unsigned int.