Initial version of a tool to output all orig/*/copyright files
This is outputted in the debian format.

Note that on it's own, this script won't do because we don't
actually ship the orig/ files with debian. Instead, the system
ships the data/ and maps/ directories, so we would need to find
the files generated from the orig/ files, and makr *those*
as the files in the copyright

Or we just ship the orig/ files with debian, and have a script
to regenrate it by running blender, gimp, etc.