Scale unit speeds to reasonable levels
The main player speed of 6.5 m/s is pretty fast. This
is 23.4 km/h which is about as fast as you'd expect a
reasonable human to go. The fastest ever recorded speed
is double that at 12.27 m/s by Usain Bolt; I guess we
could put the player speed up a bit if the game feels
too slow.

Zombies have been slowed to match