Update player model to 'MoCap Man' from Mixamo
Mixamo is a set of characters and animations provided free of
charge and royalty free by Adobe. This is a staging model for some
additional features such as multiple animations and embedded
textures.

One issue is that I don't have a copy of this model in the "orig"
directory, so I cannot use the build tools with this model. I'm
not sure that is a problem; I guess if changes to the model are
needed then it could be converted into a .blend file for that.